move cmd-list from lib/config/ to lib/utils/
