+++ lib/completion.js
@@ -34,7 +34,7 @@
 const npm = require('./npm.js')
 const { types, shorthands, defaults } = require('./utils/config.js')
 const deref = require('./utils/deref-command.js')
-const { aliases, cmdList, plumbing } = require('./config/cmd-list.js')
+const { aliases, cmdList, plumbing } = require('./utils/cmd-list.js')
 const aliasNames = Object.keys(aliases)
 const fullList = cmdList.concat(aliasNames).filter(c => !plumbing.includes(c))
 const nopt = require('nopt')
