+++ lib/help.js
@@ -13,7 +13,7 @@ var log = require('npmlog')
 var openUrl = require('./utils/open-url')
 var glob = require('glob')
 var didYouMean = require('./utils/did-you-mean')
-var cmdList = require('./config/cmd-list').cmdList
+var cmdList = require('./utils/cmd-list').cmdList
 var commands = cmdList
 var output = require('./utils/output.js')
