+++ lib/utils/cmd-list.js
