+++ test/lib/utils/cmd-list.js
@@ -1,4 +1,4 @@
 const t = require('tap')
-const cmdList = require('../../../lib/config/cmd-list.js')
+const cmdList = require('../../../lib/utils/cmd-list.js')
 // just snapshot it so we are made aware if it changes unexpectedly
 t.matchSnapshot(cmdList)
